+++ src/obr/OpenFOAM/case.py
@@ -333,7 +333,7 @@ class OpenFOAMCase(BlockMesh):
             )
 
         method = args["method"]
-        numberSubDomains = int(args.get("numberOfSubdomains", 0))
+        numberSubDomains = int(float(args.get("numberOfSubdomains", 0)))
         if method == "simple":
             if not numberSubDomains:
                 coeffs = [int(i) for i in args["coeffs"]]