+++ src/obr/cli.py
@@ -261,6 +261,9 @@ def run(ctx: click.Context, **kwargs):
 )
 @click.option("-c", "--config", required=True, help="Path to configuration file.")
 @click.option(
+    "-e", "--env", is_flag=True, help="Shows required environment variables and exits."
+)
+@click.option(
     "-t",
     "--tasks",
     default=-1,
@@ -291,6 +294,9 @@ def init(ctx: click.Context, **kwargs):
 
     config_str = read_yaml(kwargs)
     config_str = config_str.replace("\n\n", "\n")
+    if kwargs.get("env"):
+        sys.exit(0)
+
     config = yaml.safe_load(config_str)
 
     project = OpenFOAMProject.init_project(path=ws_fold)
@@ -431,6 +437,91 @@ def apply(ctx: click.Context, **kwargs):
         progress=True,
         np=1,
     )
+    logger.success("Successfully applied")
+
+
+@cli.command()
+@click.option("-c", "--config", required=True, help="Path to configuration file.")
+@click.option(
+    "--filter",
+    type=str,
+    multiple=True,
+    default=[],
+    help=(
+        "Pass a <key><predicate><value> value pair per occurrence of --filter."
+        " Predicates include ==, !=, <=, <, >=, >. For instance, obr submit --filter"
+        ' "solver==pisoFoam"'
+    ),
+)
+@click.pass_context
+def postProcess(ctx: click.Context, **kwargs):
+    from Owls.parser.LogFile import LogFile, transportEqn, customMatcher
+    from obr.core.core import get_latest_log
+    from .core.queries import build_filter_query
+    from copy import deepcopy
+    import json
+
+    def convert_to_numbers(df):
+        """convert all columns to float if they dont have Name in it"""
+        return df.astype({col: "float" for col in df.columns if not "Name" in col})
+
+    project, filtered_jobs = cli_cmd_setup(kwargs)
+
+    config_str = read_yaml(kwargs)
+    config_str = config_str.replace("\n\n", "\n")
+    config = yaml.safe_load(config_str)
+
+    d = config["postProcess"]
+
+    matcher = {"transpEqn": lambda args: transportEqn(**args)}
+    matcher_args = {"transpEqn": ["name"]}
+    matcher_regex = {}
+
+    for m in d["matcher"]:
+        matcher[m["name"]] = lambda args, regex: customMatcher(
+            args["name"], regex.format(**args)
+        )
+        matcher_args[m["name"]] = deepcopy(m["args"])
+        matcher_regex[m["name"]] = deepcopy(m["regexp"])
+
+    queries: list[Query] = build_filter_query(d["queries"])
+    query_results = project.query(jobs=filtered_jobs, query=queries)
+
+    records = []
+    for job in filtered_jobs:
+        record = {}
+        log = get_latest_log(job)
+        if not log:
+            continue
+        log_path = Path(job.path) / "case" / log
+
+        record = query_results[job.id]
+        record["jobid"] = job.id
+        for l in d["log"]:
+            try:
+                matcher_name = l["matcher"]
+                pass_args = {
+                    k: v for k, v in zip(matcher_args[matcher_name], l["args"])
+                }
+                if m_regex := matcher_regex.get(matcher_name):
+                    m = matcher[matcher_name](pass_args, matcher_regex[matcher_name])
+                else:
+                    m = matcher[matcher_name](pass_args)
+
+                log_file_parser = LogFile(log_path, matcher=[m])
+                df = convert_to_numbers(log_file_parser.parse_to_df())
+                for col in df.columns:
+                    try:
+                        record[col] = df.iloc[1:][col].mean()
+                    except:
+                        pass
+            except Exception as e:
+                print(e)
+        if record:
+            records.append(record)
+
+    with open("postpro.json", "w") as f:
+        json.dump(records, f)
     logger.success("Successfully applied")